\name{smoothtest-package}
\alias{smoothtest-package}
\alias{smoothtests}
\docType{package}
\title{Smooth test package}
\description{This package contains functions for performing smooth tests of goodness of fit and for computing orthonormal polynomials.\cr

Smooth tests with a fixed order are performed by \code{\link{smooth.test}}, and the data-driven versions by \code{\link{ddsmooth.test}}. The methods are described in Rayner et al. (2009).\cr

The orthonormal polynomials are computed with the function \code{\link{orth.poly}}. The generalised Emerson recurrence relations of Rayner et al. (2008) are used for this purpose.}
\details{\tabular{ll}{
Package: \tab smoothtest\cr
Type: \tab Package\cr
Version: \tab 1.1.2\cr
Date: \tab 2009-01-19\cr
License: \tab GPL version 2 or newer\cr}}
\references{Rayner, J.C.W., Thas, O. and De Boeck B. (2008). A Generalized Emerson Recurrence Relation. Aust.N.Z.J.Stat.50(3), 235-240 \cr
Rayner, J.C.W., Thas, O. and Best, D.J. (2009). Smooth Tests of Goodness-of-fit. Wiley, Singapore.}
\author{Olivier Thas and Bert De Boeck\cr

Maintainer: \email{Bert.DeBoeck@UGent.be}\cr}
\seealso{\code{\link[<pkg>:<pkg>-package]{<pkg>}},\code{\link{smooth.test}},\code{\link{ddsmooth.test}}, \code{link{orth.poly}}}
\keyword{package}
